rename: react-query, redux, axios folder structure modify
